Fix flaky test
This test was trying to assert that we didn't do a reparse by comparing
the green nodes of the two trees. We didn't hold onto the root of the
old tree, so it was possible for the root to fall away if we were
testing recoverable trees. If the root had fallen away, it meant
that asking for the old root would recover it as well, which would
result in deserializing the roots twice and the green nodes being
different.

The fix is to ensure the root doesn't fall away.