Do project reference conversion when applying a batch
We convert metadata references to project references if the output paths
match the referenced paths. Thus if we have a solution with a bunch of
metadata references, and then we're told a project has a new output
path, we have to go looking to see if we can convert those metadata
references to project references as well. To do that conversion, we
have to acquire the global workspace lock, since any other projects
that are also adding or removing metadata references at the same
time need to be all coordinated.

We also support the concept of "batching" when updating a project,
which means a project system can make a bunch of changes at once but
we'll raise only a single WorkspaceChanged event once those changes are
all done; we also try to ensure that we aren't acquiring the global
workspace lock until the batch is actually being applied; that ensures
we can do as much in parallel at once. We also recently added the
ability for the batch to be applied with an async API, so if there's
another project applying changes at once we aren't going to starve the
thread pool.

There was one bug, however: when you started a batch, and then changed
the output path of the project, that updating of the output path should
have waited until the batch was applied. We'd update the property in the
workspace correctly, but we immediately acquired the global workspace
lock trying to look for other projects to update with project
references. Functionally this was mostly fine (ignoring there's a small
window in time where we converted to project references even if a
project didn't have the updated output path), but it created a huge
performance issue: we would be acquiring the global workspace lock
synchronously, even if the caller was trying to operate in a batch
that they were going to apply asynchronously. If the user got
particularly unlucky, they might have a bunch of projects all being
loaded or reloaded at once, and they might starve off their thread pool.

The fix, conceptually, is simple: defer the conversion to project
references until the global lock is actually held once we are applying
the batch change. This resulted in a bit of churn; when we apply a batch
we have a little helper type called a SolutionChangeAccumulator that we
pass around that tracks the new Solution object, but also tracks what
the resulting "kind" should be for the solution change. Thus this commit
required this set of changes:

1. Convert the methods that did the conversion to/from project
   references to operate on a SolutionChangeAccumulator instead of
   directly reading CurrentSolution and directly setting the output.
2. Update the ChangeProjectOutputPath and ChangeProjectProperty
   functions to now operate with SolutionChangeAccumulators.
3. Update the project removal path to also now create a
   SolutionChangeAccumulator.
4. General refactorings (extracting/merging) of some of our core
   helpers where we apply changes in the end.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1473249