Consolidate calls to SubscribeExternalErrorDiagnosticUpdateSourceToSolutionBuildEvents
By calling this once when we already are on the UI thread when a
project is created, we're able to avoid calling it in a bunch of places
which was requiring UI threads in less-intuitive places. It's also
not clear to me why this is even marked as needing the UI thread at all,
but we'll address that independently.