Correctly handle a cached solution having frozen generated  documents
When we are synchronizing solutions, the previous assumption in this
code was incorrect -- we absolutely can end up with a frozen generated
document in CurrentSolution, since the remote workspace caches things
there. We're going to just run with that and consider that normal, so
we'll fix up the code to support it cleanly.

Fixes https://github.com/dotnet/roslyn/issues/57082
Fixes https://github.com/dotnet/roslyn/issues/56998