Show generator-produced diagnostics in the error list
This adds support for showing generator-produced diagnostics in the
error list. The implementation generally follows how we handle
diagnostics for files that we were unable to read in the IDE -- there is
a placeholder analyzer that represents the diagnostics in our various
state sets.

The one interesting thing this changes a bit is our handling of the
full-solution analysis flag. While testing this (and seeing how the file
load diagnostics worked), I noticed that we had an early bail in
"analyze project" that meant two things:

Unless you had FSA on, there was no way to actually view file load
diagnostics; this seemed odd since you then might have diagnostics in
your error list about missing types, but you wouldn't see the root
problem.  If you turned FSA back off, nothing would ever run to make the
FSA-produced diagnostics go away, short of you manually opening files or
something.  Thus I am changing the FSA handling a bit: if FSA is off,
rather than skip everything we are only skipping the analysis now; this
ensures that we are clearing out old/stale diagnostics, and we have an
opportunity to still show the generator and file load diagnostics that
we may have on hand. We only still produce those for other files (or
things in the dependency graph) for files that are open though -- the
key here is unlike regular diagnostics, these are diagnostics we know
that came from the very production of the compilation itself. If the
file is open, we're already going to have compilations being created for
it (indeed, at least for open file diagnostics!), so we're not creating
any extra work than we already had.

Fixes https://github.com/dotnet/roslyn/issues/49664
Fixes https://github.com/dotnet/roslyn/issues/58435
Fixes https://github.com/dotnet/roslyn/issues/54098