Fix up handling for partially written generics
We have some code to handle partially written generics where it could be
a type but the compiler is still parsing it as an expression based on
context. This code needed updating for the new unsigned shift operator,
but while doing and adding tests I discovered it was broken if you
called it while you were in the middle of what the compiler parsed as a
shift operator. This fixes that issue too.