Simplify RegexCharClass.IsWordChar
Roslyn contains a copy of RegexCharClass code from several years ago; it's been overhauled since, and the implementation of IsWordChar (now renamed to IsBoundaryWordChar to better reflect what it is) was improved; since Roslyn doesn't actually need all of the character set processing code, which IsBoundaryWordChar no longer uses, we can remove a bunch of unnecessary stuff.  This will now be simpler and faster.