Update dependency versions for 4.2.0 publishing
Updates our nuget versions for 4.2.0, so that all published binaries will depend on publicly-available non-prerelease packages. As part of this, I had to break add multiple versions of MSBuild, as Microsoft.VisualStudio.Shell.15.0 is now advertising their dependency on MSBuild 17.1.0. There was always an indirect dependency there, but it was not included in their nuspec.