Add a hack to remove output from the Razor generator if suppressed
The Razor generator is controlled by a flag that lives in an
.editorconfig file; in the IDE we generally don't run the generator
and instead use the design-time files added through the legacy
IDynamicFileInfo API. When we're doing Hot Reload we then
remove those legacy files and remove the .editorconfig file that is
supposed to disable the generator; for the Hot Reload pass we then are
running the generator. This is done in the CompileTimeSolutionProvider.

https://github.com/dotnet/sdk/pull/24928 introduced an issue where even
though the Razor generator is being told to not run, it still runs
anyways. As a tactical fix rather than reverting that PR, for Visual
Studio 17.3 Preview 2 we are going to do a hack here which is to rip out
generated files. In discussion, simply reverting the SDK change
was problematic because it would reintroduce some notable performance
issues that it was intended to fix in the first place. Fixing the
generator directly is also possible (and we'll be doing a follow-up
PR to try doing that), but is also a fairly risky change given the
amount of testing that was required to test the PR in the first place.
This we think is the least-risky option we have, even though it itself
is touching a fairly sensitive area of our codebase.

The intent is to remove this hack by Preview 3 with a properly tested
fix in the generator.