Add Move Static Member To Existing Type (#61519)
Ports #56131 and updates slightly. I did my best to keep the initial intent of the PR because @Soreloser2 did a great job with it :)

From the original PR:

Adds support for moving static members to existing types.
Searches for types that:

Exist within the same project as the original type
Are in the same or nested within the same namespace as the original type
Example: if we have Ns.Class1 as our original type, then Ns.Class2 and Ns.ExtraNs.Class3 will show, but DifferentNs.Class4 will not
Are the same type kind as the original type
Will not show interfaces, only shows modules if triggered on a module for VB)
Have the same or more restrictive static-ness
If the original type is a static class, then it will only show static classes. If the original type is non-static, it will show both static and non-static types (that fit the other criteria)
Types will be displayed with their full namespace qualification, any generic type parameters, and their file name (in case of partially declared types)

The user can select a pre-defined type using the dropdown menu, or type in a new, separate non-conflicting type to create a new type in a new file. New files will have an icon as well as a preview of the new type name, similar to the previous iteration of this feature. Currently existing types do not have any icon, but this is open to change based on feedback.

Co-authored-by: Soren Dahl <43689321+Soreloser2@users.noreply.github.com>
Co-authored-by: Joey Robichaud <jorobich@microsoft.com>