Fix race condition in FixAll tests based off TrackNodes functionality
Recently we enabled FixAll functionality for a bunch of code refactorings. We have been seen intermittent failures in CI for the FixAll tests added for this. I was able to reproduce the failures locally quite consistently by running the test under a loop with thousands of iterations. Upon investigation it was found that when the test fails, the [underlying CWT](https://github.com/dotnet/roslyn/blob/73de4e98d3020c72a2eef479d267357c15332d80/src/Compilers/Core/Portable/Syntax/SyntaxNodeExtensions_Tracking.cs#L18) used by TrackNodes functionality seems to have cleared the entries for nodes requested to be tracked by this FixAll operations. Adding a ToArray invocation to the tracked nodes to realize the nodes into an array seems to ensure that this doesn't happen and the test passes consistently after this change.

Ideally, we would have expected the `SyntaxEditor.OriginalRoot` to keep all the nodes in the original tree alive, but that doesn't seem to be the case. Long term, we may want to consider adding a new TrackNodes API that does the tracking using dictionaries instead of CWTs and returns a disposable tracker object, such that the caller explicitly controls the lifetime of the tracking.