Fix deadlock during the startup of diagnostics
The first time some component tries to report diagnostics, the
diagnostic service starts up and was synchronously jumping to the
UI thread to fetch the IErrorList service from Visual Studio. This was
a deadlock risk because the start of the diagnostic service could happen
from any thread, but we were blocking on the availability of the UI
thread for that to actually work.

This deadlock risk turned out to be true for at least one unlucky
user as they hit a specific case where the UI thread was blocked
trying to acquire the workspace lock. The workspace lock was held by
another thread that was also trying to acquire a lock around the
creation of an analyzer. A third thread was holding that analyzer lock
and since it was trying to report diagnostics, deadlocked here since
the UI thread was not available to get everything moving again.

Looking at the code, I don't see any reason the code in StartListening
needs to be happening synchronously; it appears to be correctly looking
at the current state of CurrentSolution to see what projects might
have been added before it first noticed. Making this properly async
is probably the best improvement we can make, because not only does
it fix this one specific user's unlucky deadlock, it also fixes any
other cases where other components are also pushing diagnostics and
might find ourselves in a loop. It also generally brings this bit of
code into compliance with the VS threading rules.

Fixes https://developercommunity.visualstudio.com/t/Deadlock-Upon-Opening-Visual-Studio/10037282