Separate out connection failures (#62136)
* Separate out connection failures

This changes the compiler to have a specific failure when it cannot
connect to the server. This lets us treat it separately from cases where
the server connected succeeded but the server rejected the request.

The reason for this change is cannot connect failures are not
necessarily an indicator of a code issue. A fully functional build can
hit connection issues for a variety of reasons. For example if the
server is too busy and can't respond to a connection event in time.

Having a separate response means we can handle the failures differently
in bootstrap builds. The bootstrap build will not fail now until it sees
more than three of these. That should reduce the false positive rate we
see in our infra.

* PR feedback