Introduce a new API Solution.WithCachedSourceGeneratorState
For Razor, we only run their generator on forked solutions that are
thrown away; this API gives us a way to reuse that cached state in other
forked solutions, since otherwise there's no way to reuse the cached
state.

This API can be deleted once we get rid of the concept of compile-time
versus design-time compilations that we currently have for Razor.