Use the prior cached state for the Razor generator in later runs
The CompileTimeSolutionProvider forks the main Workspace's Solution
snapshot and temporarily enables the Razor generator for that forked
version. When a later edit were to happen, we'd end up throwing out that
snapshot and re-forking the Solution again, this meant that any cached
state from the Razor generator gets thrown out for later runs. This
uses the new API just added to try to properly cache the Razor generator
output from the prior run.