When we're removing the last project, fire SolutionRemoved
We have a fair bit of code that assumed it could look for a
WorkspaceChangeKind of SolutionRemoved or SolutionCleared to indicate
when we've closed the solution and can drop caches and such. Except we
never actually fired anything like that. This fires SolutionRemoved on
the removal of the last project.

We had a test that asserted that the solution persistence service would
continue to return the solution path even once the solution was closed;
this fixes that test to assert that it's no longer available.