Delete our SetCallstackIfEmpty helper
When an exception object is created but unthrown, the call stack is
simply null. This helper tries to "fix" that, but when it does
so the exception's StackTrace will not be a full stack trace, but
simply the frames that were unwound during the throw...which is just
this method.

Since this doesn't do anything useful, deleting it means the VS
Telemetry library will be able to do it's own handling around missing
stack traces, which it actually has code to deal with.