Fix analyzer action state tracking for partial analysis
Fixes #63205

While executing analyzers for a tree sub-span on CompilationWithAnalyzers, we were not adding to the set of processed nodes/operations after all the actions have been executed for it. This meant that all actions would be re-executed for this node/operation when analyzer diagnostics are subsequently queried on the same CompilationWithAnalyzers instance.