Expose IsGeneratedCode boolean property on diagnostic analyzer contexts
Closes #7578

Based on the recent feature requests in allowing more fine grained generated code analysis, we have decided to initially expose a boolean IsGeneratedCode property on each of the analysis contexts. Note that this flag indicates if the callback symbol/node/operation/tree is considered generated code or not based on compiler's internal heuristics for analyzer execution. This flag helps the analyzers that only want to analyze certain kinds of generated code to narrow down their custom fine grained generated code analysis to only those callbacks that have this flag set to true.