Change cancel behavior for rename (#63600)
Rename cancel behavior depending on caret position change in the editor is fragile. Instead, it's better to just depend on focus events for the UI and dismiss as needed. 

This also allows Cancel to be called multiple times for an InlineRenameSession, and just bail if it's already dismissed

Fixes the new rename behavior in vsvim