Tweak selection algorithm when filter text is lowercase only
Before we'd always consider case-insensitive matches equally good as case-sensitive ones, and select the one with the highested MatchPriority, but this turns out to be problematic in certain sceanrios (e.g. https://github.com/dotnet/roslyn/issues/63922)

To fix this issue, we now only consider case-insensitive match better then case-sensitive one if it's "MatchPriority.Preselect" or the other item it compares to is "MatchPriority.Deprioritize".

For all the other values of MatchPriority, it'd be only take into account during item selection if two items have same text match result.