Add some tests to verify that frozen documents work
Investigating https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1467404,
there was one possibility that we had a bug somewhere where we
incorrectly used a wrong project state when there was a series of
edits and a compilation is stale. Since we didn't really have any
tests that validated this currently, this adds tests for that scenario.