Only write out inVs properties when appropriate
LSIF edges nodes can use "inV" properties to indicate one-to-one edges,
or "inVs" (plural) to indicate one-to-many edges. I previously cheated
and used "inVs" everywhere when we wrote out JSON, and it just happened
that there was only ever a single vertex in the list when it was
supposed to be one-to-one. Some import tools were OK with this, but
validation tools still flag this as spec violation, so let's do it
properly.

Fixes https://github.com/dotnet/roslyn/issues/63841