Ensure Project.WithAnalyzerReferences deals with value-equal references
When the list of analyzer references change for a project, we produce
a new list of AnalyzerFileReferences and pass that into
Project.WithAnalyzerReferences. The implementation of that method
figures out which references are added or removed, and also uses
those changes to update the list of generators being held by the
generator.

This seems innocent enough, except that the AnalyzerFileReferences
here implement value equality: two references are equal if they have
the same file path. These references however will not return the same
generator instances, because each reference does it's own loading
and caching of that list. Thus, when we computed the list of analyzers
that are added or removed, we won't see analyzer references that are
equal, and won't update the generator driver. Later code that expects
those to be in sync will then start throwing exceptions.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1655835