If the last generator is removed that was generating, correctly update
If the last generator was removed, and it was generating trees, we would
have potentially left that tree in the Compilation that was returned.
I believe this would have been a transient issue -- any later change
to the project would have created a new CompilationTracker with an
InProgress state; the code that processes an InProgress state into
the final state would have correctly seen we no longer had a generator
and would have dropped the old compilation at that point.