Remove the static ConditionalWeakTable for per-project CompilationWithAnalyzers in the OOP process
Fixes [AB#1549502](https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1549502)

Currently, we maintain a static CWT to cache the per-project CompilationWithAnalyzers instance for analyzer execution in the OOP process. The primary goal is to share the same CompilationWithAnalyzers instance for document analysis for active and open documents in the project. Before this CWT was introduced, we used a fresh CompilationWithAnalayzers instance (and hence a fresh compilation instance) for every open document analyzed in background, which led to a huge performance overhead in creating compilations. However, it seems that the CWT approach is leading to the opposite problem of us holding onto too much of memory when the number of projects/open documents are large. Analysis of [AB#1549502](https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1549502) shows more than 100+ instances of CompilationWithAnalyzers instances in the CWT.

This PR switches the OOP DiagnosticComputer to an in-between caching approach by only caching the CompilationWithAnalyzers instance for the last project to be analyzed. This allows us to re-use the CompilationWithAnalyzers instance for open document analysis while also not holding onto it in memory for too long. Solution crawler background analysis already ensures that it goes project by project for computing open document diagnostics, and the LSP-based diagnostics host that eventually replaces the solution crawler should take a similar approach to maximize the re-use of this cache.

Even if the approach taken in this PR turns out to be insufficient to give optimal performance characteristics, we definitely want to move away from the CWT based approach as that is definitely problematic for large solutions with many open documents as shown by the above perf bug.